+++ packages/pos-client/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from "./validator.js";
+export * from "./misc.js";