+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @walletconnect/utils
 
+## 2.23.1
+
+### Patch Changes
+
+- [#7085](https://github.com/WalletConnect/walletconnect-monorepo/pull/7085) [`41c8ebd`](https://github.com/WalletConnect/walletconnect-monorepo/commit/41c8ebd19739aa8a346a7c054a75320243860a77) Thanks [@ganchoradkov](https://github.com/ganchoradkov)! - Updated `@walletconnect/logger` to `3.0.1`
+
+- Updated dependencies [[`41c8ebd`](https://github.com/WalletConnect/walletconnect-monorepo/commit/41c8ebd19739aa8a346a7c054a75320243860a77)]:
+  - @walletconnect/types@2.23.1
+
 ## 2.23.0
 
 ### Minor Changes